--- ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/DeleteCommunityCommentDto.kt
@@ -1,8 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto
-
-class DeleteCommunityCommentDto {
-
-    data class Response(
-        val id: Long,
-    )
-}
