+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/comment/CreateCommunityCommentDto.kt
@@ -1,4 +1,4 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
 
 class CreateCommunityCommentDto {
