+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/comment/DeleteCommunityCommentDto.kt
@@ -1,0 +1,1 @@
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment
