+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/CreateCommunityPostCommand.kt
@@ -1,9 +1,8 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.posts
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
 
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 
-class UpdateCommunityPostCommand(
-    val id: Long,
+class CreateCommunityPostCommand(
     val title: String,
     val content: String,
     val categoryType: CommunityCategoryType,
