+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/DeleteCommunityPostCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
+
+class DeleteCommunityPostCommand(
+    val id: Long,
+) {
+}
