+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/GetCommunityPostCommand.kt
@@ -1,0 +1,6 @@
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
+
+class GetCommunityPostCommand(
+    val id: Long,
+) {
+}
