+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/GetCommunityPostDto.kt
@@ -1,25 +1,14 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.posts
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
 
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 import java.time.LocalDateTime
 
-class SearchCommunityPostDto {
-
-    data class Request(
-        val categoryType: CommunityCategoryType?,
-        val subwayLine: String?,
-        val title: String?,
-        val content: String?,
-        val hashTag: String?
-    )
+class GetCommunityPostDto {
 
     data class Response(
-        val posts: List<CommunityPost>
-    )
-
-    data class CommunityPost(
         val id: Long,
         val title: String,
+        val content: String,
         val categoryType: CommunityCategoryType,
         val views: Int,
         val likes: Int,
