+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/post/UpdateCommunityPostCommand.kt
@@ -1,8 +1,9 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.posts
+package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.post
 
 import backend.team.ahachul_backend.api.community.domain.model.CommunityCategoryType
 
-class CreateCommunityPostCommand(
+class UpdateCommunityPostCommand(
+    val id: Long,
     val title: String,
     val content: String,
     val categoryType: CommunityCategoryType,
