--- ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/posts/DeleteCommunityPostCommand.kt
@@ -1,6 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.posts
-
-class DeleteCommunityPostCommand(
-    val id: Long,
-) {
-}
