--- ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/adapter/web/in/dto/posts/GetCommunityPostCommand.kt
@@ -1,6 +1,0 @@
-package backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.posts
-
-class GetCommunityPostCommand(
-    val id: Long,
-) {
-}
