+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/application/port/in/CommunityCommentUseCase.kt
@@ -1,6 +1,9 @@
 package backend.team.ahachul_backend.api.community.application.port.`in`
 
-import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.*
+import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment.CreateCommunityCommentDto
+import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment.DeleteCommunityCommentDto
+import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment.GetCommunityCommentsDto
+import backend.team.ahachul_backend.api.community.adapter.web.`in`.dto.comment.UpdateCommunityCommentDto
 
 interface CommunityCommentUseCase {
