+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/community/domain/model/CommunityPostType.kt
@@ -1,0 +1,5 @@
+package backend.team.ahachul_backend.api.community.domain.model
+
+enum class CommunityPostType {
+    CREATED, DELETED
+}
