+++ ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/api/member/domain/model/MemberStatusType.kt
@@ -1,5 +1,5 @@
 package backend.team.ahachul_backend.api.member.domain.model
 
-enum class MemberStatus {
+enum class MemberStatusType {
     ACTIVE, SUSPENDED, DELETE
 }
