--- ahachul_backend/src/main/kotlin/backend/team/ahachul_backend/common/model/CommunityPostType.kt
@@ -1,5 +1,0 @@
-package backend.team.ahachul_backend.common.model
-
-enum class CommunityPostType {
-    CREATED, DELETED
-}
